requirements.txt
amqp==5.0.6
asgiref==3.2.10
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dataclasses==0.8
Django==3.0.5
django-rename-app==0.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
djongo==1.3.4
idna==2.10
importlib-metadata==4.0.1
kombu==5.0.2
mongoengine==0.23.1
prompt-toolkit==3.0.18
PyJWT==1.7.1
pymongo==3.11.4
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.16.0
sqlparse==0.2.4
typing-extensions==3.10.0.0
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
